test : 테스트 변수명 정리 및 Fixtures 분리
