+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -1,31 +1,23 @@
-package nextstep.subway.path.acceptance.application;
+package nextstep.subway.path.acceptance;
 
 import static nextstep.subway.line.acceptance.step.LineAcceptanceStep.지하철_노선_등록되어_있음;
 import static nextstep.subway.line.acceptance.step.LineSectionAcceptanceStep.지하철_노선에_지하철역_등록_요청;
+import static nextstep.subway.path.acceptance.step.PathAcceptanceStep.*;
 import static nextstep.subway.station.step.StationAcceptanceStep.지하철역_등록되어_있음;
-import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 
 @DisplayName("지하철 경로 조회")
-public class PathAcceptanceTest extends AcceptanceTest {
+class PathAcceptanceTest extends AcceptanceTest {
 
     private LineResponse 신분당선;
     private LineResponse 이호선;
@@ -49,7 +41,6 @@ public class PathAcceptanceTest extends AcceptanceTest {
         천호역 = 지하철역_등록되어_있음("천호").as(StationResponse.class);
         강동구청 = 지하철역_등록되어_있음("강동구청").as(StationResponse.class);
 
-        //String lineName, String color,Long upStationId, Long downStationId, int distance
         신분당선 = 지하철_노선_등록되어_있음("신분당선", "bg-red-600", 강남역.getId(), 양재역.getId(), 10).as(
             LineResponse.class);
         이호선 = 지하철_노선_등록되어_있음("이호선", "bg-red-600", 교대역.getId(), 강남역.getId(), 10).as(
@@ -70,7 +61,9 @@ public class PathAcceptanceTest extends AcceptanceTest {
 
         // then
         최단경로_조회_됨(response, Arrays.asList(강남역.getId(), 양재역.getId(), 남부터미널역.getId()));
+        최단경로_조회_길이_계산됨(response, 12);
     }
+
 
     @Test
     void 같은_역_경로_조회_실패() {
@@ -90,29 +83,4 @@ public class PathAcceptanceTest extends AcceptanceTest {
         경로_조회_실패됨(response);
     }
 
-    private void 경로_조회_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    public static ExtractableResponse<Response> 경로_조회(Long source, Long target) {
-        Map<String, Long> query = new HashMap<>();
-        query.put("source", source);
-        query.put("target", target);
-        return RestAssured
-            .given().log().all()
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .queryParams(query)
-            .when().get("/paths")
-            .then().log().all().
-            extract();
-    }
-
-    private void 최단경로_조회_됨(ExtractableResponse<Response> response, List<Long> expected) {
-        PathResponse line = response.as(PathResponse.class);
-        List<Long> stationIds = line.getStations().stream()
-            .map(StationResponse::getId)
-            .collect(Collectors.toList());
-
-        assertThat(stationIds).containsExactlyElementsOf(expected);
-    }
 }
